3d rotating cube

+++ src/tests/Test.hpp
@@ -14,7 +14,7 @@ namespace Tests {
         Test();
         virtual ~Test();
 
-        virtual void update(float deltaTime) {};
+        virtual void update(double deltaTime) {};
         virtual void render(const Renderer &renderer) = 0;
         virtual void renderUI() {};
     };
