+++ src/tests/TestClearColor.hpp
@@ -12,7 +12,7 @@ namespace Tests {
         TestClearColor();
         ~TestClearColor();
 
-        virtual void update(float deltaTime) override;
+        virtual void update(double deltaTime) override;
         virtual void render(const Renderer &renderer) override;
         virtual void renderUI() override;
     };
